+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+### Feature git
+
 - Remove need for context in Sentry.init for Android ([#117](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/117))
 
 ## 0.2.1