+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - Remove need for context in Sentry.init for Android ([#117](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/117))
+- Bump Java SDK from v6.14.0 to v6.32.0 ([#130](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/130))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6320)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.14.0...6.32.0)
 
 ## 0.2.1