+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Add sample & trace rate configuration ([#144](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/144))
 - Remove need for context in Sentry.init for Android ([#117](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/117))
 
 ### Dependencies