+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 - Bump Java SDK from v6.14.0 to v6.33.1 ([#139](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/139))
   - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#6331)
-  - [diff](https://github.com/getsentry/sentry-java/compare/6.32.0...6.33.1)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.14.0...6.33.1)
 
 ## 0.2.1